Consolidate Adornable class logic into new BaseAdornableNode
